creating services, repositories etc

+++ src/main/java/com/guides/repositories/GuideRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.guides.models.Guide;
 
 public interface GuideRepository extends JpaRepository<Guide, Integer> {
-	List<Guide> findByInvitedUsersId(Long userId);
+	List<Guide> findByInvitedUsersId(int userId);
 }
